+++ R/learndown_learnr.R
@@ -120,6 +120,8 @@ record_learnr <- function(tutorial_id, tutorial_version, user_id, event, data) {
     session     = "", # Should we use this?
     date        = format(Sys.time(),
       format = "%Y-%m-%d %H:%M:%OS6", tz = "GMT"),
+    type        = "learnr",
+    id          = "",
     app         = tutorial_id,
     version     = version,
     user        = user_id,